# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: AKJ Travels API
  version: 1.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/auth:
    post:
      summary: Registrar un nuevo usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserDTO' }
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/auth/login:
    post:
      summary: Iniciar sesión
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginDTO' }
      responses:
        '200':
          description: Token emitido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/forgot-password:
    post:
      summary: Solicitar código de restablecimiento
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestResetPasswordDTO' }
      responses:
        '200':
          description: Código enviado por email
          content:
            text/plain:
              schema: { type: string }
  /api/auth/reset-password:
    patch:
      summary: Restablecer contraseña con código
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordDTO' }
      responses:
        '200':
          description: Contraseña cambiada
          content:
            text/plain:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }

  ############################
  # USERS
  ############################
...
  /api/bookings/{id}:
    delete:
      summary: Cancelar/eliminar reserva
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: Reserva eliminada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
  /api/bookings/{id}/confirm:
    post:
      summary: Confirmar reserva (flujo de host)
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: Reserva confirmada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }

  ############################
  # FAVORITES
  ############################
  /api/favorites/{placeId}:
    post:
      summary: Marcar lugar como favorito (mi usuario)
      tags: [Favorites]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Marcado como favorito
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }

    delete:
      summary: Quitar de favoritos (mi usuario)
      tags: [Favorites]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Eliminado de favoritos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
  /api/favorites/me:
    get:
      summary: Listar mis favoritos (paginado)
      tags: [Favorites]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Página de favoritos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponsePlacePage' }
  /api/favorites/me/{placeId}:
    get:
      summary: ¿Este lugar es mi favorito?
      tags: [Favorites]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: true/false
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseBoolean' }
  /api/favorites/count/{placeId}:
    get:
      summary: Número de usuarios que marcaron favorito este lugar
      tags: [Favorites]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Conteo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseLong' }

  ############################
  # COMMENTS (responder)
  ############################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      in: query
      name: size
      required: false
      schema: { type: integer, minimum: 1, default: 10 }
  responses:
    BadRequest:
      description: Petición inválida
    Unauthorized:
      description: No autorizado
    Conflict:
      description: Conflicto
...
  schemas:
...
    ResponseString:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { type: string, example: "ok" }
    ResponseBoolean:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { type: boolean, example: true }
    ResponseLong:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { type: integer, format: int64, example: 12 }
    ResponseToken:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { $ref: '#/components/schemas/TokenDTO' }
...
    ResponsePlacePage:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:
          type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/PlaceEntitySummary' }
            totalElements: { type: integer, format: int64, example: 23 }
            totalPages:    { type: integer, example: 3 }
            size:          { type: integer, example: 10 }
            number:        { type: integer, example: 0 }
...
    LoginDTO:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, minLength: 7, maxLength: 20 }
    TokenDTO:
      type: object
      properties:
        token: { type: string }
    CreateUserDTO:
      type: object
      required: [name, surname, email, phone, birthDate, country, password, role]
      properties:
        name:      { type: string }
        surname:   { type: string }
        email:     { type: string, format: email }
        phone:     { type: string }
        birthDate: { type: string, format: date }
        country:   { type: string }
        photoUrl:  { type: string, nullable: true }
        password:  { type: string, minLength: 8, description: "Debe incluir mayúscula y dígito" }
        role:      { $ref: '#/components/schemas/Role' }
...
    RequestResetPasswordDTO:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordDTO:
      type: object
      required: [code, email, newPassword]
      properties:
        code:        { type: string }
        email:       { type: string, format: email }
        newPassword: { type: string, minLength: 6 }
...
    PlaceEntitySummary:
      type: object
      description: Resumen de entidad Place para favoritos/listas
      properties:
        id:            { type: string }
        title:         { type: string }
        price:         { type: number, format: double }
        city:          { type: string }
        averageRating: { type: number, format: double }
        photoUrl:      { type: string, nullable: true }

    # DTOs BOOKINGS/COMMENTS
...
    Role:
      type: string
      enum: [USER, HOST, ADMIN]
...